changed URL for API & config application.properties

+++ src/main/java/com/spring/ecommerce/persistence/model/Product.java
@@ -21,6 +21,6 @@ public class Product {
     private String imageURL;
     private String description;
     private Double price;
-    private Integer ratting;
+    private Double ratting;
     private String evaluatting;
 }
